Began working on the enemies.
System to manage enemies has been worked on, currently all types of
enemies can be loade, waves cannot yet be loaded but EnemyManager has
been worked on and it has partial logic frame finished (wave spawning
part) and it can already find a proper place to spawn enemies on and
find path for them to reach their destination.
In addition to that, documentation has been moved to folders
Documentation and some typos in the comments have been fixed.